backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai-whisper
setuptools-rust

groq
yt-dlp==2023.12.30
moviepy==1.0.3
python-multipart==0.0.9
requests==2.31.0
ffmpeg-python==0.2.0
